Rename tab components + fix some linting errors
