+++ packages/ui/src/lib/sidebar/TabPanel.svelte
@@ -1,7 +1,7 @@
 <script lang='ts'>
 	import type { Tab } from './sidebarState'
 
-	import { isWideView, selectedTabId } from './sidebarState'
+	import { isWideView } from './sidebarState'
 
 	export let tabs: Tab[] = []
 
@@ -9,7 +9,7 @@
 	$: classes = $isWideView ? 'h-full w-52' : 'w-full h-52'
 </script>
 
-<div class="bg-core-grey-800 {classes}">
+<div role="tabpanel" class="bg-core-grey-800 {classes}">
 	<slot>
 		<div class="text-white h-full w-full pt-8 flex justify-center">Todo</div>
 	</slot>
